Added link to OONI / testyourinter.net to the widget
Note: I’m not sure if there was a reason why links to share and donate
were not typical links and are handled with js. I added
testyourinter.net as a normal <a> link. If there was a reason to do it
the former way, please fix. Also, it would be good to add an SVG image
icon here, perhaps a phone icon, or an app icon.